+++ libft/src/ft_putnbr_fd.c
@@ -6,13 +6,13 @@
 /*   By: dpiza <dpiza@student.42sp.org.br>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/05 22:27:36 by dpiza             #+#    #+#             */
-/*   Updated: 2021/06/08 16:01:21 by dpiza            ###   ########.fr       */
+/*   Updated: 2021/08/20 16:27:06 by dpiza            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_putnbr_fd(int n, int fd)
+void	ft_putnbr_fd(long long n, int fd)
 {
 	if (n == -2147483648)
 	{
